+++ dto/backchannel_authentication_complete_response.go
@@ -71,4 +71,10 @@ type BackchannelAuthenticationCompleteResponse struct {
 
 	//
 	Resources []string `json:"resources,omitempty"`
+
+	//
+	ServiceAttributes []Pair `json:"serviceAttributes,omitempty"`
+
+	//
+	ClientAttributes []Pair `json:"clientAttributes,omitempty"`
 }